requirements.txt
fastapi~=0.110.0
uvicorn
pydantic
pytest
httpx
telegram
pyyaml~=6.0.1
python-telegram-bot==13.7
requests~=2.31.0
psycopg2-binary

APScheduler==3.6.3
async-timeout==4.0.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
configparser==6.0.0
idna==3.6
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
tornado==6.4
tzlocal==5.2
urllib3==1.26.18
